this:红客 充值 + 充值记录 ok next:微信登陆  完成注册

+++ HoKol/HokolAlbum/src/main/java/com/hokol/activity/StarDynamicActivity.java
@@ -53,7 +53,7 @@ public class StarDynamicActivity extends BaseAppCompatActivity
 	private boolean isPraised;
 
 	// 用户 coin数目
-	private int userCoin;
+	private float userCoin;
 
 	public static void actionStart(Context context, String dynamicId)
 	{
@@ -205,7 +205,7 @@ public class StarDynamicActivity extends BaseAppCompatActivity
 				LogFileUtil.v("发送礼物成功");
 				SDKManager.toast("礼物赠送成功");
 
-				int restCoinNum = userCoin - giftCoinNum;
+				float restCoinNum = userCoin - giftCoinNum;
 				AppStateManager.getInstance().updateKeyUserCoinNum(StarDynamicActivity.this, restCoinNum);
 			}
 		});
